tests failing but the preparsing error works
