+++ tdrs-backend/tdpservice/parsers/parse.py
@@ -6,6 +6,7 @@ import itertools
 import logging
 from .models import ParserErrorCategoryChoices, ParserError
 from . import schema_defs, validators, util, row_schema
+from .schema_defs.util import get_section_reference, get_program_model
 
 logger = logging.getLogger(__name__)
 
@@ -58,7 +59,7 @@ def parse_datafile(datafile):
 
     section_is_valid, section_error = validators.validate_header_section_matches_submission(
         datafile,
-        util.get_section_reference(program_type, section),
+        get_section_reference(program_type, section),
         util.make_generate_parser_error(datafile, 1)
     )
 
@@ -297,4 +298,4 @@ def manager_parse_line(line, schema_manager, generate_error, is_encrypted=False)
 def get_schema_manager(line, section, program_type):
     """Return the appropriate schema for the line."""
     line_type = line[0:2]
-    return util.get_program_model(program_type, section, line_type)
+    return get_program_model(program_type, section, line_type)
