+++ tdrs-backend/tdpservice/parsers/schema_defs/tanf/t1.py
@@ -12,6 +12,7 @@ t1 = SchemaManager(
             document=TANF_T1DataSubmissionDocument(),
             preparsing_validators=[
                 validators.hasLength(156),
+                validators.field_year_month_with_header_year_quarter(),
             ],
             postparsing_validators=[
                 validators.if_then_validator(
