+++ tdrs-backend/tdpservice/parsers/test/test_parse.py
@@ -12,6 +12,7 @@ from tdpservice.search_indexes import documents
 from .factories import DataFileSummaryFactory
 from tdpservice.data_files.models import DataFile
 from .. import schema_defs, aggregates, util
+from ..schema_defs.util import get_section_reference, get_program_models
 
 import logging
 
@@ -847,7 +848,7 @@ def test_get_schema_options(dfs):
     assert schema == schema_defs.tanf.t1
 
     # get model
-    models = util.get_program_models('TAN', 'A')
+    models = get_program_models('TAN', 'A')
     assert models == {
         'T1': schema_defs.tanf.t1,
         'T2': schema_defs.tanf.t2,
@@ -857,7 +858,7 @@ def test_get_schema_options(dfs):
     model = util.get_program_model('TAN', 'A', 'T1')
     assert model == schema_defs.tanf.t1
     # get section
-    section = util.get_section_reference('TAN', 'C')
+    section = get_section_reference('TAN', 'C')
     assert section == DataFile.Section.CLOSED_CASE_DATA
 
     # from datafile:
