+++ tdrs-backend/tdpservice/parsers/validators.py
@@ -1,7 +1,7 @@
 """Generic parser validator functions for use in schema definitions."""
 
 from .models import ParserErrorCategoryChoices
-from .util import fiscal_to_calendar
+from .util import fiscal_to_calendar, year_month_to_year_quarter
 from datetime import date
 import logging
 
@@ -23,10 +23,25 @@ def value_is_empty(value, length, extra_vals={}):
 
 # higher order validator func
 
+class change_func_str:
+    class FNMagic:
+        def __init__(self,fn,fn_name):
+            self.fn = fn 
+            self.fn_name = fn_name
+        def __call__(self,*args,**kwargs):
+            return self.fn(*args,**kwargs)
+        def __str__(self):
+            return self.fn_name
+    def __init__(self,name):
+       self.fn_name = name
+    def __call__(self,fn):
+        return self.FNMagic(fn,self.fn_name)
 
 def make_validator(validator_func, error_func):
     """Return a function accepting a value input and returning (bool, string) to represent validation state."""
-    def validator(value):
+
+    #@change_func_str(validator_func.__str__())
+    def validator(value, instance=None):
         try:
             if validator_func(value):
                 return (True, None)
@@ -157,6 +172,40 @@ def sumIsEqual(condition_field, sum_fields=[]):
 
     return lambda value: sumIsEqualFunc(value)
 
+"""
+def field_year_month_with_header_year_quarter():
+    def validate_reporting_month_year_fields_with_header(line, df_quarter, df_year):
+
+        print('++++++++++++++++++', f"{line}")
+
+        # get reporting month year from header
+        field_year, field_quarter = year_month_to_year_quarter(f"{field_month_year}")
+        file_calendar_year, file_calendar_qtr = fiscal_to_calendar(df_year, f"{df_quarter}")
+        return (True, None) if str(file_calendar_year) == str(field_year) and file_calendar_qtr == field_quarter else (
+            False, f"Reporting month year {field_month_year} does not match file reporting year:{df_year}, quarter:{df_quarter}.",
+            )
+
+    return lambda value, df_quarter, df_year: validate_reporting_month_year_fields_with_header(value, df_quarter, df_year)
+"""
+
+def field_year_month_with_header_year_quarter():
+    def validate_reporting_month_year_fields_with_header(line, row_schema_instance):
+
+        field_month_year = row_schema_instance.get_field_values_by_names(line, ['RPT_MONTH_YEAR']).get('RPT_MONTH_YEAR')
+        df_quarter = row_schema_instance.datafile.quarter
+        df_year = row_schema_instance.datafile.year
+
+        # get reporting month year from header
+        field_year, field_quarter = year_month_to_year_quarter(f"{field_month_year}")
+        file_calendar_year, file_calendar_qtr = fiscal_to_calendar(df_year, f"{df_quarter}")
+        print('_______________str(file_calendar_year) == str(field_year):', str(file_calendar_year) == str(field_year))
+        print('_______________file_calendar_qtr == field_quarter:', file_calendar_qtr == field_quarter)
+        return (True, None) if str(file_calendar_year) == str(field_year) and file_calendar_qtr == field_quarter else (
+            False, f"Reporting month year {field_month_year} does not match file reporting year:{df_year}, quarter:{df_quarter}.",
+            )
+
+    return lambda value, row_schema_instance: validate_reporting_month_year_fields_with_header(value, row_schema_instance)
+
 
 def sumIsLarger(fields, val):
     """Validate that the sum of the fields is larger than val."""
@@ -240,8 +289,8 @@ def contains(substring):
 def startsWith(substring):
     """Validate that string value starts with the given substring param."""
     return make_validator(
-        lambda value: value.startswith(substring),
-        lambda value: f"{value} does not start with {substring}.",
+        lambda value, instance=None: value.startswith(substring),
+        lambda value, instance=None: f"{value} does not start with {substring}.",
     )
